Time: 0 ms (100.00%) | Memory: 18.5 MB (42.17%) - LeetSync

+++ 1071-binary-prefix-divisible-by-5/binary-prefix-divisible-by-5.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     vector<bool> prefixesDivBy5(vector<int>& nums) {
-        long long ans = 0;
+        int ans = 0;
         int n = nums.size();
         vector<bool> res(n);
         for(int i = 0; i < n; ++i){
